requirements.txt
tensorflow
openai-whisper
numpy
opencv-python
librosa
matplotlib
pandas
scikit-learn
tqdm
torch